abstract
Sharing bicycle has attracted the most sharply increased attention throughout the world in recent years.
Its rapid development has increased transportation convenience. However, abandoned bicycles have
caused a waste of resources and environmental damage. Additionally, the negative effects will exceed the
positive impact that sharing bicycle creates if these abandoned bicycles are handled improperly. This
paper proposes a two-stage method to study the recycling of abandoned bicycles, including the selection
of the third-party logistics suppliers (3 PLs) and the recycling network design. In the first stage, interval
type-2 fuzzy sets (IT2FSs) and Vlse Kriterjumska Optimizacija I Kompromisno Resenje (VIKOR) are used
to select the 3 PLs. The dynamic time series, individual regrets, and group efficiency are all taken into
account. In the second stage, From the perspective of maximizing overall benefits, a mathematical model
considering the purchase cost of raw materials, transportation costs, recovery costs, and environmental
benefits is established to design a recycling network. Finally, the applicability of the proposed method is
numerically demonstrated by solving a real case, and key conclusions are that the recycling mode with
the participation of the 3 PLs in sharing bicycle system is the most reasonable and feasible. However, in
this mode, the selection criteria of the 3 PLs are different from other industries, which need to be given
enough attention by the sharing bicycle operators. Additionally, in the recycling network design,
considering the maximum overall interests of the three parties, the factory, the 3 PLs, and the operators,
must coincide with the social philosophy of sustainable development. The results are of great guiding
significance to the recycling of sharing bicycles.